+++ heron-pulsar-client/src/main/java/org/apache/heron/pulsar/PulsarBoltConfiguration.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pulsar.storm;
+package org.apache.heron.pulsar;
 
 import java.util.Objects;
