+++ heron-pulsar-client/src/main/javadoc/overview.html
@@ -21,9 +21,9 @@
 <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN">
 <HTML>
   <HEAD>
-    <TITLE>Pulsar Storm API Overview</TITLE>
+    <TITLE>Pulsar Heron API Overview</TITLE>
   </HEAD>
   <BODY>
-    The Pulsar Storm API is a proprietary messaging API.
+    The Pulsar Heron API is a proprietary messaging API.
   </BODY>
 </HTML>
